feat(Line): 지하철 노선을 의미하는 클래스
