+++ src/main/java/subway/domain/subway/line/Line.java
@@ -1,0 +1,37 @@
+package subway.domain.subway.line;
+
+import java.util.Objects;
+
+public class Line {
+    private final String name;
+
+    public Line(String name) {
+        this.name = name;
+    }
+
+    public String getName() {
+        return name;
+    }
+
+    @Override
+    public boolean equals(Object o) {
+        if (this == o) {
+            return true;
+        }
+        if (o == null || getClass() != o.getClass()) {
+            return false;
+        }
+        Line line = (Line) o;
+        return Objects.equals(name, line.name);
+    }
+
+    @Override
+    public int hashCode() {
+        return Objects.hash(name);
+    }
+
+    @Override
+    public String toString() {
+        return "Line{" + name + '\'' + '}';
+    }
+}
